Move Timer component to components folder
