+++ src/components/Timer/Timer.style.ts
