+++ src/components/Timer/Timer.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import { setTime } from "./TimerSlice";
-import { useAppDispatch, useAppSelector } from "../App/hooks";
+import { useAppDispatch, useAppSelector } from "../../App/hooks";
 import { addMilliseconds, differenceInMilliseconds, format } from "date-fns";
 
 const Timer = () => {
